DataManager - Redistribución del acceso a datos en memoria
Se ha redistribuido el acceso de datos a memoria en una única clase. Además se le han realizado controles de overflow  a las variables del sistema.

Arreglo de bug menor en checkFPS

+++ S-PAC-3 MAN/Assets/Scripts/Quest/Quest.cs
@@ -7,9 +7,9 @@ public class Quest{
     [XmlAttribute("Description")]
     public string description;
     [XmlAttribute("RequiredAmount")]
-    public int requiredAmount;
+    public float requiredAmount;
     [XmlAttribute("Reward")]
-    public int reward;
+    public float reward;
     [XmlAttribute("Type")]
     public int type;
 }
